Add files via upload
This code can be run on MATLAB version 2013.The file path should be copied and paste it in the MATLAB. And using the Sender and Receiver command we can run this project. The encrypted image block are been shared by using AWS S3 cloud service and it is been shared with the Receiver.